Refactor rate-limited warning logic into a separate module
This commit extracts the rate-limiting logic for dropped records from `FemtoStreamHandler` into a new `RateLimiter` struct in its own module.

This addresses issue #83 by reducing code duplication and improving maintainability.
