+++ rust_extension/Cargo.toml
@@ -7,6 +7,9 @@ edition = "2021"
 name = "_femtologging_rs"
 crate-type = ["cdylib", "rlib"]
 
+[build-dependencies]
+pyo3-build-config = "0.25.1"
+
 [dependencies]
 pyo3 = { version = ">=0.25.1,<0.26.0", default-features = false, features = ["macros", "auto-initialize"] }
 crossbeam-channel = "0.5.15"
