+++ rust_extension/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    pyo3_build_config::use_pyo3_cfgs();
+}
